+++ src/resource/types.h
@@ -30,6 +30,7 @@ struct MeshAnimationResource;
 struct PackageResource;
 struct PhysicsConfigResource;
 struct RenderConfigResource;
+struct StatConfigResource;
 struct ShaderResource;
 struct ShapeResource;
 struct SoundResource;
@@ -68,6 +69,7 @@ struct Platform
 #define RESOURCE_TYPE_PACKAGE          STRING_ID_64("package",          UINT64_C(0xad9c6d9ed1e5e77a))
 #define RESOURCE_TYPE_PHYSICS_CONFIG   STRING_ID_64("physics_config",   UINT64_C(0x72e3cc03787a11a1))
 #define RESOURCE_TYPE_RENDER_CONFIG    STRING_ID_64("render_config",    UINT64_C(0x27862fe24795319c))
+#define RESOURCE_TYPE_STAT_CONFIG      STRING_ID_64("stat_config",      UINT64_C(0xb6be805d51b0e770))
 #define RESOURCE_TYPE_SCRIPT           STRING_ID_64("lua",              UINT64_C(0xa14e8dfa2cd117e2))
 #define RESOURCE_TYPE_SHADER           STRING_ID_64("shader",           UINT64_C(0xcce8d5b5f5ae333f))
 #define RESOURCE_TYPE_SOUND            STRING_ID_64("sound",            UINT64_C(0x90641b51c98b7aac))
@@ -92,6 +94,7 @@ struct Platform
 #define RESOURCE_VERSION_PACKAGE          RESOURCE_VERSION(11)
 #define RESOURCE_VERSION_PHYSICS_CONFIG   RESOURCE_VERSION(5)
 #define RESOURCE_VERSION_RENDER_CONFIG    RESOURCE_VERSION(8)
+#define RESOURCE_VERSION_STAT_CONFIG      RESOURCE_VERSION(1)
 #define RESOURCE_VERSION_SCRIPT           RESOURCE_VERSION(4)
 #define RESOURCE_VERSION_SHADER           RESOURCE_VERSION(18)
 #define RESOURCE_VERSION_SOUND            RESOURCE_VERSION(2)