+++ docs/changelog.rst
@@ -24,6 +24,7 @@ Changelog
 * Tools: added an option under Preferences -> Viewport to choose the rendering backend used by the editor runtimes.
 * Tools: improved rendering performance of selected objects.
 * Tools: improved performance when moving many units in the Level Editor.
+* Tools: added a new ``stat`` console command to show performance statistics.
 * Runtime: improved performance when rendering worlds containing many units.
 * Runtime: improved performance of sun shadows rendering.
 * Runtime: Linux: added a new ``--display-server`` option to select between X11 and Wayland backends. X11 remains the default.